+++ dom/media/mediasink/moz.build
@@ -8,6 +8,7 @@ UNIFIED_SOURCES += [
     'AudioSinkWrapper.cpp',
     'DecodedAudioDataSink.cpp',
     'DecodedStream.cpp',
+    'OutputStreamManager.cpp',
     'VideoSink.cpp',
 ]
